Apply drop and keep modifiers to rolls
