+++ src/evaluate/mod.rs
@@ -1,3 +1,4 @@
 pub mod dice_roll;
 pub mod expression;
 pub mod group_rolls;
+pub mod roll;
